Create/delete notification after like/unlike post

+++ app/Livewire/LikeButton.php
@@ -2,7 +2,10 @@
 
 namespace App\Livewire;
 
+use App\Enums\NotificationContent;
+use App\Models\Notification;
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class LikeButton extends Component
@@ -18,11 +21,28 @@ class LikeButton extends Component
         $user = auth()->user();
 
         $hasLiked = $user->likes()->where('post_id', $this->post->id)->exists();
-        if($hasLiked){
+        if ($hasLiked) {
             $user->likes()->detach($this->post);
+
+            $notification = Notification::where('receiver_id', $this->post->user_id)
+                                         ->where('sender_id', Auth::user()->id)
+                                         ->where('type', 'like')
+                                         ->first();
+
+            if ($notification) {
+                $notification->delete();
+            }
         } else {
             $user->likes()->attach($this->post);
+        
+            Notification::create([
+                'receiver_id' => $this->post->user_id,
+                'sender_id' => Auth::user()->id,
+                'type' => 'like',
+                'content' => NotificationContent::TYPES['liked']
+            ]);
         }
+        
     }
     public function render()
     {
